+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v2.24.2
 
 * added DeferredValue
+* satella.coding.Context are considered experimental